cmd/pi-strings-go: implement utf16 support using x/text/encoding/unicode
